+++ s/data/base64url.ts
@@ -3,25 +3,34 @@ import {Bytes} from "./bytes.js"
 import {Base64} from "./base64.js"
 
 export const Base64url = {
-
-	string(bytes: Uint8Array) {
-		return Base64.string(bytes)
+	fromBytes(bytes: Uint8Array) {
+		return Base64.fromBytes(bytes)
 			.replace(/\+/g, "-")
 			.replace(/\//g, "_")
 			.replace(/=+$/g, "")
 	},
 
-	bytes(string: string) {
+	toBytes(string: string) {
 		let b64 = string
 			.replace(/-/g, "+")
 			.replace(/_/g, "/")
 		if (b64.length % 4 !== 0)
 			b64 = b64.padEnd(b64.length + (4 - b64.length % 4) % 4, "=")
-		return Base64.bytes(b64)
+		return Base64.toBytes(b64)
 	},
 
 	random(count = 32) {
-		return this.string(Bytes.random(count))
+		return this.fromBytes(Bytes.random(count))
+	},
+
+	/** @deprecated renamed to "fromBytes" */
+	string(bytes: Uint8Array) {
+		return Base64url.fromBytes(bytes)
+	},
+
+	/** @deprecated renamed to "toBytes" */
+	bytes(string: string) {
+		return Base64url.toBytes(string)
 	},
 }